Add service targets
Currently, the way we handle startup is a hodgepodge of mess adapted
from dinit's example and Chimera Linux. It's about time we standardize
this, and given that it's likely that dinit (and I personally) would
probably gravitate to dinit-chimera, I can also refactor a lot of
things (a.k.a. fix my horrible mistakes), hopefully we can sort our
dependency tree to make it more readable.